refactor: 패키지 구조 변경 및 테스트 클래스 수정
패키지 구조를 변경하여 도메인 클래스를 보다 명확하게 구분하였습니다.
- Message, ChatRoom, RoomRole, RoomType 클래스의 패키지를 수정하였습니다.
- RoomParticipantTest 클래스의 내용을 수정하여 RoomParticipant 도메인 테스트로 변경하였습니다.
