+++ qootalk-server/module-domain/src/main/java/com/lrchan/qootalk/domain/chat/room/RoomType.java
@@ -1,4 +1,4 @@
-package com.lrchan.qootalk.domain.chat;
+package com.lrchan.qootalk.domain.chat.room;
 
 public enum RoomType {
     DIRECT,
