+++ qootalk-server/module-domain/src/test/java/com/lrchan/qootalk/domain/chat/room/ChatRoomTest.java
@@ -1,4 +1,4 @@
-package com.lrchan.qootalk.domain.chat;
+package com.lrchan.qootalk.domain.chat.room;
 
 import static org.assertj.core.api.Assertions.assertThat;
